Random seed population tests (#122)
