+++ pyEpiabm/pyEpiabm/tests/test_unit/test_routine/test_toy_population_config.py
@@ -57,50 +57,15 @@ class TestPopConfig(TestPyEpiabm):
         patch_log.assert_called_once_with("ValueError in ToyPopulation"
                                           + "Factory.make_pop()")
 
-    def summarise_pop(self, pop):
-        # Returns lists of cell and microcell wise populations
-        # Not a testing function, but used in test below
+    @patch("numpy.random.seed")
+    @patch("random.seed")
+    def test_random_seed_param(self, mock_random, mock_np_random, n=42):
+        pop_params = {"population_size": 100, "cell_number": 10,
+                      "microcell_number": 10, "population_seed": n}
+        ToyPopulationFactory.make_pop(pop_params)
 
-        pop_cells = []  # List of populations in each cell of population
-        pop_microcells = []  # List of populations in each microcell
-        for cell in pop.cells:
-            pop_cells.append(len(cell.persons))
-            for microcell in cell.microcells:
-                pop_microcells.append(len(microcell.persons))
-        return pop_cells, pop_microcells
-
-    @parameterized.expand([(random.randint(1000, 10000),
-                            random.randint(5, 10),
-                            random.randint(2, 10),
-                            random.randint(1, 100))
-                          for _ in range(numReps)])
-    def test_pop_seed(self, pop_size, cell_number, microcell_number, seed):
-        """Tests for when the population is implemented by default with
-        no households. Parameters are assigned at random.
-        """
-        # Define parameters for population generation
-        pop_params = {"population_size": pop_size, "cell_number": cell_number,
-                      "microcell_number": microcell_number,
-                      "population_seed": seed}
-
-        # Create two identical populations with the same seed
-        seed_pop = ToyPopulationFactory.make_pop(pop_params)
-        comp_pop = ToyPopulationFactory.make_pop(pop_params)
-
-        self.assertEqual(str(seed_pop), str(comp_pop))
-
-        seed_cells, seed_microcells = self.summarise_pop(seed_pop)
-        comp_cells, comp_microcells = self.summarise_pop(comp_pop)
-        self.assertEqual(seed_cells, comp_cells)
-        self.assertEqual(seed_microcells, comp_microcells)
-
-        # Also compare to a population with a different seed
-        pop_params["population_seed"] = seed + 1  # Change seed of population
-        diff_pop = ToyPopulationFactory().make_pop(pop_params)
-
-        diff_cells, diff_microcells = self.summarise_pop(diff_pop)
-        self.assertNotEqual(seed_cells, diff_cells)
-        self.assertNotEqual(seed_microcells, diff_microcells)
+        mock_random.assert_called_once_with(n)
+        mock_np_random.assert_called_once_with(n)
 
     @parameterized.expand([(random.randint(1000, 10000) * numReps,
                             random.randint(1, 10) * numReps,
