+++ pyEpiabm/pyEpiabm/tests/test_unit/test_sweep/test_host_progression_sweep.py
@@ -29,6 +29,26 @@ class TestHostProgressionSweep(TestPyEpiabm):
             "prob_icu_to_icurecov": 0.5,
             "prob_icu_to_death": 0.5
         }
+        self.mock_inf_prog = np.array(
+            [0.487464241, 1, 1.229764827, 1.312453175,
+             1.307955665, 1.251658756, 1.166040358,
+             1.065716869, 0.960199498, 0.855580145,
+             0.755628835, 0.662534099, 0.577412896,
+             0.500665739, 0.432225141, 0.371729322,
+             0.318643018, 0.272340645, 0.232162632,
+             0.19745264, 0.167581252, 0.141960133,
+             0.120049578, 0.101361532, 0.085459603,
+             0.071957123, 0.060514046, 0.050833195,
+             0.04265624, 0.035759641, 0.029950735,
+             0.025064045, 0.02095788, 0.017511251,
+             0.014621091, 0.012199802, 0.010173075,
+             0.008477992, 0.007061366, 0.005878301,
+             0.00489096, 0.004067488, 0.003381102,
+             0.00280931, 0.002333237, 0.001937064,
+             0.001607543, 0.001333589, 0.001105933,
+             0.00091683, 0.000759816, 0.000629496,
+             0.000521372, 0.000431695, 0.000357344,
+             0.000295719, 0.000244659])
 
         # Create population that will be used to test all
         #  methods except update status
@@ -211,26 +231,7 @@ class TestHostProgressionSweep(TestPyEpiabm):
         with mock.patch('pyEpiabm.Parameters.instance') as mock_param:
             mock_param.return_value.time_steps_per_day = 1
             mock_param.return_value.asympt_infect_period = 14
-            mock_param.return_value.infectiousness_prof =\
-                np.array([0.487464241, 1, 1.229764827, 1.312453175,
-                          1.307955665, 1.251658756, 1.166040358,
-                          1.065716869, 0.960199498, 0.855580145,
-                          0.755628835, 0.662534099, 0.577412896,
-                          0.500665739, 0.432225141, 0.371729322,
-                          0.318643018, 0.272340645, 0.232162632,
-                          0.19745264, 0.167581252, 0.141960133,
-                          0.120049578, 0.101361532, 0.085459603,
-                          0.071957123, 0.060514046, 0.050833195,
-                          0.04265624, 0.035759641, 0.029950735,
-                          0.025064045, 0.02095788, 0.017511251,
-                          0.014621091, 0.012199802, 0.010173075,
-                          0.008477992, 0.007061366, 0.005878301,
-                          0.00489096, 0.004067488, 0.003381102,
-                          0.00280931, 0.002333237, 0.001937064,
-                          0.001607543, 0.001333589, 0.001105933,
-                          0.00091683, 0.000759816, 0.000629496,
-                          0.000521372, 0.000431695, 0.000357344,
-                          0.000295719, 0.000244659])
+            mock_param.return_value.infectiousness_prof = self.mock_inf_prog
             # Parameters to determine where the tail starts, has to be the same
             # as for the parameters called in HostProgressionSweep.
             infectious_period = pe.Parameters.instance().asympt_infect_period
@@ -260,26 +261,7 @@ class TestHostProgressionSweep(TestPyEpiabm):
             # steps per day:
             mock_param.return_value.time_steps_per_day = 100
             mock_param.return_value.asympt_infect_period = 14
-            mock_param.return_value.infectiousness_prof =\
-                np.array([0.487464241, 1, 1.229764827, 1.312453175,
-                          1.307955665, 1.251658756, 1.166040358,
-                          1.065716869, 0.960199498, 0.855580145,
-                          0.755628835, 0.662534099, 0.577412896,
-                          0.500665739, 0.432225141, 0.371729322,
-                          0.318643018, 0.272340645, 0.232162632,
-                          0.19745264, 0.167581252, 0.141960133,
-                          0.120049578, 0.101361532, 0.085459603,
-                          0.071957123, 0.060514046, 0.050833195,
-                          0.04265624, 0.035759641, 0.029950735,
-                          0.025064045, 0.02095788, 0.017511251,
-                          0.014621091, 0.012199802, 0.010173075,
-                          0.008477992, 0.007061366, 0.005878301,
-                          0.00489096, 0.004067488, 0.003381102,
-                          0.00280931, 0.002333237, 0.001937064,
-                          0.001607543, 0.001333589, 0.001105933,
-                          0.00091683, 0.000759816, 0.000629496,
-                          0.000521372, 0.000431695, 0.000357344,
-                          0.000295719, 0.000244659])
+            mock_param.return_value.infectiousness_prof = self.mock_inf_prog
             infectious_period = pe.Parameters.instance().asympt_infect_period
             model_time_step = 1 / pe.Parameters.instance().time_steps_per_day
             num_infectious_ts = int(np.ceil(infectious_period /
@@ -312,25 +294,7 @@ class TestHostProgressionSweep(TestPyEpiabm):
                 mock_param.return_value.time_steps_per_day = 1
                 mock_param.return_value.asympt_infect_period = 14
                 mock_param.return_value.infectiousness_prof =\
-                    np.array([0.487464241, 1, 1.229764827, 1.312453175,
-                              1.307955665, 1.251658756, 1.166040358,
-                              1.065716869, 0.960199498, 0.855580145,
-                              0.755628835, 0.662534099, 0.577412896,
-                              0.500665739, 0.432225141, 0.371729322,
-                              0.318643018, 0.272340645, 0.232162632,
-                              0.19745264, 0.167581252, 0.141960133,
-                              0.120049578, 0.101361532, 0.085459603,
-                              0.071957123, 0.060514046, 0.050833195,
-                              0.04265624, 0.035759641, 0.029950735,
-                              0.025064045, 0.02095788, 0.017511251,
-                              0.014621091, 0.012199802, 0.010173075,
-                              0.008477992, 0.007061366, 0.005878301,
-                              0.00489096, 0.004067488, 0.003381102,
-                              0.00280931, 0.002333237, 0.001937064,
-                              0.001607543, 0.001333589, 0.001105933,
-                              0.00091683, 0.000759816, 0.000629496,
-                              0.000521372, 0.000431695, 0.000357344,
-                              0.000295719, 0.000244659])
+                    self.mock_inf_prog
                 num_infectious_ts = int(np.ceil(pe.Parameters.instance().
                                                 asympt_infect_period
                                                 * pe.Parameters.instance().
@@ -396,26 +360,7 @@ class TestHostProgressionSweep(TestPyEpiabm):
             # Parameters used in update infectiousness for both time steps
             # length:
             mock_param.return_value.asympt_infect_period = 14
-            mock_param.return_value.infectiousness_prof =\
-                np.array([0.487464241, 1, 1.229764827, 1.312453175,
-                          1.307955665, 1.251658756, 1.166040358,
-                          1.065716869, 0.960199498, 0.855580145,
-                          0.755628835, 0.662534099, 0.577412896,
-                          0.500665739, 0.432225141, 0.371729322,
-                          0.318643018, 0.272340645, 0.232162632,
-                          0.19745264, 0.167581252, 0.141960133,
-                          0.120049578, 0.101361532, 0.085459603,
-                          0.071957123, 0.060514046, 0.050833195,
-                          0.04265624, 0.035759641, 0.029950735,
-                          0.025064045, 0.02095788, 0.017511251,
-                          0.014621091, 0.012199802, 0.010173075,
-                          0.008477992, 0.007061366, 0.005878301,
-                          0.00489096, 0.004067488, 0.003381102,
-                          0.00280931, 0.002333237, 0.001937064,
-                          0.001607543, 0.001333589, 0.001105933,
-                          0.00091683, 0.000759816, 0.000629496,
-                          0.000521372, 0.000431695, 0.000357344,
-                          0.000295719, 0.000244659])
+            mock_param.return_value.infectiousness_prof = self.mock_inf_prog
             infectious_period = pe.Parameters.instance().asympt_infect_period
 
             # Generating the list of infectiousness values for when we have 1
@@ -486,26 +431,7 @@ class TestHostProgressionSweep(TestPyEpiabm):
         mock_param.return_value.model_time_step = 1
         mock_param.return_value.asympt_infect_period = 14
         mock_param.return_value.sympt_infectiousness = 1.5
-        mock_param.return_value.infectiousness_prof =\
-            np.array([0.487464241, 1, 1.229764827, 1.312453175,
-                      1.307955665, 1.251658756, 1.166040358,
-                      1.065716869, 0.960199498, 0.855580145,
-                      0.755628835, 0.662534099, 0.577412896,
-                      0.500665739, 0.432225141, 0.371729322,
-                      0.318643018, 0.272340645, 0.232162632,
-                      0.19745264, 0.167581252, 0.141960133,
-                      0.120049578, 0.101361532, 0.085459603,
-                      0.071957123, 0.060514046, 0.050833195,
-                      0.04265624, 0.035759641, 0.029950735,
-                      0.025064045, 0.02095788, 0.017511251,
-                      0.014621091, 0.012199802, 0.010173075,
-                      0.008477992, 0.007061366, 0.005878301,
-                      0.00489096, 0.004067488, 0.003381102,
-                      0.00280931, 0.002333237, 0.001937064,
-                      0.001607543, 0.001333589, 0.001105933,
-                      0.00091683, 0.000759816, 0.000629496,
-                      0.000521372, 0.000431695, 0.000357344,
-                      0.000295719, 0.000244659])
+        mock_param.return_value.infectiousness_prof = self.mock_inf_prog
         # First check that people progress through the
         # infection stages correctly.
         self.person2.update_status(pe.property.InfectionStatus.Exposed)
